refactor: playGame 기능을 game-count.js에서 분리해 모듈화 기존 game-count.js가 count 기능에 더 충실할 수 있고, 새로운 game-play.js가 play 기능에 더 충실할 수 있게 변경
